Initial commit of the neurospin branch (old fff2 code).
This work is the beginning of the major merge of all the current FFF2
library developed at Neurospin:

http://www-dsv.cea.fr/neurospin/

by the team of Bertrand Thirion, Alexis Roche, Benjamin Thyreau, Gael
Varoquaux, Merlin Keller, Philippe Ciuciu and collaborators.

FFF2 is functional imaging algorithms library which we are now merging into
the NIPY core code.  Initially this code will live in a special 'neurospin'
module so that we can easily isolate the changes and work on it, though
eventually its algorithms and functionality will migrate gradually across
the entire NIPY codebase.

For current users of the FFF2 codebase, NIPY will provide (once this work is
merged into the trunk) a compatibility layer, so that upon installing NIPY,
the statement

imprort fff2

will continue to work, though now it will actually pull its functionality
from within the 'import neuroimaging' package.  This will be done to ease,
as much as possible, the transition for existing FFF2 users.

For further information on the team behind FFF2, please see:

http://www.lnao.fr/
http://www.inria.fr/recherche/equipes/parietal.en.html