Large set of changes to be able to build the entire documentation.
Note that the docs are as of now quite messy, but at least (with current
matplotlib, this will be fixed late) both HTML and PDF do build and all of
the hand-written FFF2 existing docs get built, as well as
'neuroimaging.neurospin' API docs.

The Neurospin team will then progressively update their docs.