RF - raise error for default image iteration
Images should not iterate by default because, by default, they iterate
over the first axis, which is often not what we want.  For example, the
fourth axis may well be the time axis.  To avoid confusion and make this
explicit, we raise an error, preferring the explicit ``iter_axis``
function instead.