RF: contains space check allows axes in any order
Previously contains would check if the axes for the checked object were
in the exact order of the space against which it was being checked.
That meant that, if we roll time first for an MNI space image, it would
no longer be `in` MNI space. With these changes, the image continues to
be in MNI space, but cannot - for example - generate an xyz affine
without rearrangement.