ENH: add warning for unhandled intent codes
We don't handle any nifti intent codes. We could; for example, if the
intent code is 3 ('t test') we'd expect the 4th dimension of the dataset
to be length 1 (not time), and the 5th dimension of the dataset to be 1
or 2 in length, 1 for just the t value, 2 for the t value and the
degrees of freedom for a each voxel.  We could then give the 4th
dimension in our nipy image a label like 't params'.  But maybe that's
for another day, and when we see that these codes are being used.