BF: residual outputs reversed domain and range
The residual output reversed the coordmap domain and range. Until now
the tests didn't pick this up because we allowed through XYZ output
names that didn't match expected names for NIFTI. The nifti writer does
now check for these, and tranposes the image accordingly. Now the image
is being transposed, the test needs to look for the NIFTI transposed
shape.