TEST: re-enable load / save tests for reordering
Jonathan disabled these tests when the nifti save did not reorder the
data axes according to x, y, z.  It looks like the old code left the
input axes as (x, y, z) in the first 3, but in any order.  The current
code does reorder to (x, y, z) exactly, and I adapted the tests to this.
Also, we now use the nifti ``toffset`` for time offset.