RF+TST: refactor convolve_functions interface
Considering problem in github issue 16 - how to allow for different
support and intervals for the f and g functions being convolved?

Solve by passing in the intervals for f and g.

Remove padding_t keyword because we're returning the full convolution,
and this gives us some tail from the convolution.  The user can expand
the f interval to be safer about edge effects.

Add fill=0 and kwargs argument to pass to interpolator

Because we are using step functions for the tests, very small floating
point differences can push the values to the left or right of the
threshold, causing delta t's worth of errors.  So, in the tests, use
intervals where we know the floating point result to test with.