RF: stop setup.py importing from nipy tree
It was importing from the nipy tree to see if we had data packages
installed.  The data routines since moved to nibabel, so it's more or
less as easy to check directly in the setup.py.  This also means we
don't have to import from the code tree before extensions are built, or
2to3 processing has been done.