NF: convert nipy image with 'unknown' world to nii
We previously raised an error if we had a nipy image with the 'unknown'
image space when converting to nifti, because 'unknown' cannot be stored
in nifti unless the affine is discarded.  But, if the affine of the
image to be converted is already the same as that resulting from
discarding the affine, there's no harm, and it allows us to round-trip
images without specified (``sform, qform``) affines.