RF: run local script files through Python
On Windows, we can't run the local files through the shell because they
don't have needed ``.bat`` or ``.py`` extension. On Posix, we might have
the wrong Python interpreter named on the first line - for example -
"#!/usr/bin/env python" won't necessarily work for every platform, and
might not be the Python we're interested in.  So, run the scripts
through the interpreter we're running in already, if the scripts are
from the source.