MISC: remove transparency
Using transparency is a bad idea because if you save as figures as PDFs,
the corresponding PDFs will rely on transparency in PDF which
cannot be translated to EPS (EPS does not have the notion of
transparency). As a result, publishers like Elsevier will rasterize the
figure, often in a very crude way. People keep getting bitten by this
effect. It is better to change the defaults.