DOC: updating doctests sphinx docs
Doctests are all now passing on my box, but with some extra changes in
the code, that I will submit as a separate pull request. Basically I
needed to allow objects for coordinate map coordinate types to allow the
sympy things that Jonathan is doing in the long coordinate map
discussion.