BF: make windows script wrappers more robust
The script wrappers were hard-coding the python path at the time of the
creation of the binary such as an exe or an egg.  This path is likely to be
wrong after the binary is installed somewhere.  Use some minor bat file
programming to pull the Python command out of the script directly, rather like
the ``cli.exe`` trick used by setuptools.