RF: doctest uses Term rather than symbol
Latest sympy lambdify needs the symbol on the left and the right hand
side of a lambdified function to be the same.  Hence we need 't' to be
the term 't' not any sympy symbol 't'.  At the moment we're using our
own replacement for lambdify, but this change allows us to use the stock
lambdify as of 0.7.3