BF: change in sympy sorting changes doctest output
We've changed from using ``sort`` on sympy expressions, to using
``sort(expr, key=sympy.default_sort_key)``, in order to support sympy
trunk (> 0.7.3).  Unfortunately, these produce different results when
sorting dummy variables, as in:

In [38]: a = sympy.Symbol('a')

In [39]: b = sympy.Dummy('b')

In [40]: sorted([a, b])
Out[40]: [a, _b]

In [41]: sorted([a, b], key=sympy.default_sort_key)
Out[41]: [_b, a]

I can't think of a way round this, so just change the failing doctest to
use the new order.