RF: refuse to load / convert MINC for now
Nibabel can load MINC, but now we need an API to get the axis names for
MINC, otherwise, the axes can be anywhere, but are most likely to be in
the opposite order to the ones the NIfTI needs.  Raise an error until we
have worked out how to do this right.