BF+TST: fix bug for adding and subtracting Factors
Without properly understanding the code, I previously (a8f6f44) made
Formula add and subtract return an instance of the added / subtracted
class, but that forced the assumption that the class constructor for the
subclass is the the same as Formula, which is not the case for Factor.
Fix and test.

Partially fixes gh-258.