API+RF: remove doctest skipping machinery
The doctest skipping machinery was causing testing failure with the
six.py module, because the machinery runs `getattr` on the object it is
testing which in the case of six.py, was causing various `moves` to be
imported.

As we were only using the doctest skipping machinery for one doctest,
and we've just removed that doctest, remove the machinery.