BF: fix and test sdist command
I broke setup.py sdist in the Python 3 refactor, because we were
previously using a version of numpy.distutils build_py, and the refactor
reverted (accidentally) to using the default disutils setup.py.

We also forgot to add the new position of the fff files to the
MANIFEST.in, so the sdist archives was broken.