RF: refactor cosine basis to use dct-ii function
Use dedicated function for _cosine basis.

This refactoring will require continuous times, otherwise it will raise
an error.  In any case, I believe that the values returned when times
are not continuous are not valid:

a) The current function assumes that all differences in volume times are
equal to the difference in the first two volume times, and;
b) The DCT values should (I believe) reflect the number of `dt`
intervals, not the index into the time vector, as currently.