BF: work around sympy handling of longcomplex
Sympy does not handle np.longcomplex correctly at the moment, and this
started raising an error with Sympy 1.1, released recently. It was
discarding the imaginary part before, so no real loss.  See:

https://github.com/sympy/sympy/pull/12901

for a proposal for proper handling of np.longcomplex.  In the mean time,
don't use it for testing.