Cython: declare pointers one per line
None of these declarations was incorrect, but Cython now warns about
them. The developers are worried that people will write “cdef int* a, b”
and think that “b” is an “int*” rather than an “int”. Reducing the
number of warnings from Cython makes it easier to notice warnings that
might represent real problems.